2024/day13/main_test.go
package main

import (
	_ "embed"
	"testing"
)

//go:embed example.txt
var example string

var example1 string = `Button A: X+94, Y+34
Button B: X+22, Y+67
Prize: X=8400, Y=5400`

type test struct {
	name, input string
	want        int
}

func Test_part1(t *testing.T) {
	tests := []test{
		{
			"example 1",
			example1,
			280,
		},
		{
			"full example",
			example,
			480,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part1(tt.input); got != tt.want {
				t.Errorf("part1() = %d; want %d\n", got, tt.want)
			}
		})
	}
}

func Test_part2(t *testing.T) {
	tests := []test{
		{
			"full example",
			example,
			0,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part2(tt.input); got != tt.want {
				t.Errorf("part2() = %d; want %d\n", got, tt.want)
			}
		})
	}
}
